Ajout de quelques fonctionnalités pour les scripts bash

+++ script_install.sh
@@ -7,8 +7,7 @@
 # Appel de notre bibliothèque de fonctions
 source biblio.sh
 
-ME=$USER
-if [[ -z "$ME" ]] || [[ "$ME" = "root" ]]; then
+if [[ -z "$ME" ]] || [[ $ID_ME -eq $ID_ROOT ]]; then
     ME2=$(ls -l /home | grep -E "^d.*\+" | rev | awk '{print $1}' | rev | head -n 1)
     number=$(echo "$ME2" | wc -l)
     if [[ "$number" -ne 1 ]] || [[ ! -d "/home/$ME2" ]]; then
@@ -18,39 +17,58 @@ if [[ -z "$ME" ]] || [[ "$ME" = "root" ]]; then
     fi
 fi
 
-if [[ "$(whoami)" != "root" ]]; then
+if ! is_root; then
     echo -e "${RED}Erreur : le script doit être exécuté en tant que superutilisateur (root).${NC}" 2>&1
     exit 4
 fi
 directory="/home/$ME/Téléchargements"
 cd "$directory"
+
 package_manager="apt"
 extension="deb"
-if [ "$#" -eq 2 ]; then
+ALL_PACKAGES="$1"
+if [ "$#" -eq 3 ]; then
     package_manager="$1"
     extension="$2"
+    ALL_PACKAGES="$3"
 fi
+
 if ! command -v "$package_manager" >/dev/null 2>&1; then
 	echo -e "${RED}Erreur : le système doit être basé sur Debian pour exécuter le script.${NC}" 2>&1
 	exit 3
 fi
+
+if [ ! -f "$ALL_PACKAGES" ]; then 
+    echo -e "${RED}Le fichier '$ALL_PACKAGES' n'est pas un fichier régulier valide. Terminaison.${NC}"
+    exit 1
+fi 
+
 echo -e "${YELLOW}Gestionnaire utilisé : ${NC}${GREEN}$package_manager${NC}"
 echo -e "${YELLOW}Extension standard : ${NC}${GREEN}$extension${NC}\n"
+
 echo -ne "${YELLOW}Récupération de google chrome..... ${NC}"
 wget "https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.$extension" -O "$directory/chrome.$extension" > /dev/null 2>&1
 check_cmd "google chrome"
+
 echo -ne "${YELLOW}Récupération de discord..... ${NC}"
 wget https://discordapp.com/api/download?platform=linux -O "$directory/discord.$extension" > /dev/null 2>&1
 check_cmd "discord"
+
 echo -ne "${YELLOW}Récupération de vivaldi..... ${NC}"
 wget "https://downloads.vivaldi.com/stable/vivaldi-stable_5.6.2867.58-1_amd64.$extension" -O "$directory/viv.$extension" > /dev/null 2>&1
 check_cmd "vivaldi"
+
+echo -ne "${YELLOW}Récupération de VSCodium..... ${NC}"
+wget "https://github.com/VSCodium/vscodium/releases/download/1.90.2.24171/codium_1.90.2.24171_amd64.deb" -O "$directory/codium.$extension" > /dev/null 2>&1
+check_cmd "vscodium"
+
 echo -e "\n${YELLOW}Installation des fichiers .$extension${NC}\n"
 echo "Applications non installées :" >> /home/"$ME"/error.log
 for file in "$directory"/*."$extension"; do
     chmod 755 "$file"
-    install_app $file "$package_manager"
+    install_deb "$file"
 done
+
 rm -f "$directory"/*."$extension"
 if [ "$?" -eq 0 ]; then
     echo -e "\n${GREEN}Fichiers .$extension supprimés avec succès !${NC}"
@@ -60,93 +78,56 @@ fi
 
 # Activation de php8.1
 echo -e "\n${YELLOW}Activation de PHP8.1${NC}\n"
-echo -ne "${YELLOW}Update du système..... ${NC}"
-sudo $package_manager update > /dev/null 2>&1
-check_cmd ""
+updateee "$package_manager"
+
 echo -ne "${YELLOW}Certificats pour php8.1 : ${NC}"
 $package_manager install ca-certificates apt-transport-https software-properties-common wget curl lsb-release > /dev/null 2>&1
 check_cmd "récupération des certificats pour php8.1"
+
 echo -ne "${YELLOW}Récupération de php8.1 : ${NC}"
 curl -sSL https://packages.sury.org/php/README.txt | sudo bash -x > /dev/null 2>&1
 check_cmd "récupération de php8.1"
-echo -ne "${YELLOW}Update du système : ${NC}" 
-sudo $package_manager update > /dev/null 2>&1
+
+updateee "$package_manager"
+
+echo -ne "${YELLOW}Activation de Spotify (1/2).....${NC} "
+curl -sS https://download.spotify.com/debian/pubkey_6224F9941A8AA6D1.gpg | sudo gpg --dearmor --yes -o /etc/apt/trusted.gpg.d/spotify.gpg > /dev/null 2>&1
 check_cmd ""
+
+echo -ne "${YELLOW}Activation de Spotify (2/2).....${NC} "
+echo "deb http://repository.spotify.com stable non-free" | sudo tee /etc/apt/sources.list.d/spotify.list > /dev/null 2>&1
+check_cmd ""
+
+updateee "$package_manager"
+
 # Applications à installer avec le gestionnaire de paquet défini avant
-echo -e "\n${YELLOW}Installation des fichiers avec $package_manager${NC}\n"
-applications=(
-    "gnome-tweaks"
-    "libreoffice"
-    "texlive-full"
-    "texmaker"
-    "obs-studio"
-    "audacity"
-    "htop"
-    "neofetch"
-    "nano"
-    "openjdk-17-jdk-headless"
-    "ssh"
-    "sshfs"
-    "sshpass"
-    "gcc"
-    "valgrind"
-    "clang"
-    "g++"
-    "cmake"
-    "nodejs"
-    "racket"
-    "npm"
-    "gpg"
-    "php8.1"
-    "vlc"
-    "timeshift"
-    "flatpak"
-    "python3"
-    "python3-pip"
-    "python3.11-venv"
-    "celluloid"
-    "virt-manager"
-    "git"
-    "apache2"
-    "php"
-    "libapache2-mod-php"
-    "mariadb-server"
-    "php-mysql"
-    "php-curl"
-    "php-gd"
-    "php-intl"
-    "php-json"
-    "php-mbstring"
-    "php-xml"
-    "php-zip"
-    "php-bcmath"
-)
-multiple_choices_by_propositions_array APPS "Choisissez les applications que vous voulez installer" "${applications[@]}"
-for ((i=0 ; i < ${#APPS[@]} ; i++)); do
-    install_app "${APPS[i]}" "$package_manager"
-done
-echo -ne "${YELLOW}Installation de r7rs-lib (pour racket)..... ${NC}"
-sudo -u "$ME" raco pkg install r7rs-lib > /dev/null 2>&1
-check_cmd "r7rs-lib"
+echo -e "\n${PURPLE}Installation des paquets avec $package_manager${NC}\n"
+install_all_apps_from_file "$ALL_PACKAGES"
+
+# Activation de mariadb
+echo -ne "${YELLOW}Activation de MariaDB (1/2)..... ${NC}" 
+sudo systemctl start mariadb
+check_cmd ""
+echo -ne "${YELLOW}Activation de MariaDB (2/2)..... ${NC}" 
+sudo systemctl enable mariadb
+check_cmd ""
+
 #Activation des flatpaks
-echo -ne "\n${YELLOW}Activation de Flathub : ${NC}\n" 
-flatpak remote-delete --force flathub > /dev/null > /dev/null 2>&1
+echo -ne "\n${YELLOW}Activation de Flathub..... ${NC}" 
+flatpak remote-delete --force flathub > /dev/null 2>&1
 flatpak remote-add --if-not-exists flathub https://flathub.org/repo/flathub.flatpakrepo > /dev/null 2>&1
 check_cmd "activation de flathub"
+
 # Application à installer avec flatpak
-# Evince : lecteur pdf de gnome
 echo -ne "\n${YELLOW}Installation des flatpaks : ${NC}\n"
 flatpaks=(
     "com.github.unrud.VideoDownloader"
-    "flathub com.spotify.Client"
-    "VSCodium"
-    "onlyoffice"
-    "Zotero"
-    "evince"
+    "kdenlive"
+    "flathub"
 )
 multiple_choices_by_propositions_array FLS "Choisir les flatpaks que vous voulez installer" "${flatpaks[@]}"
 for ((i=0 ; i < ${#FLS[@]} ; i++)); do
-    install_app "${FLS[i]}" "flatpak"
+    install_flatpak "${FLS[i]}"
 done
 
 # Nettoyer les dépendances inutiles
